requirements.txt
Flask
Streamlit
flask_cors
requests
numpy
pandas
python-dotenv
langchain==0.1.16
langchain-core
langchain-groq
